+++ src/games/Game.cs
@@ -1,0 +1,5 @@
+public abstract class Game {
+    public abstract string Name { get; }
+    
+    public abstract int PlayGame();
+}
