+++ src/games/testgame.cs
@@ -1,0 +1,8 @@
+public class TestGame : Game
+{
+    public override string Name => "Game";
+
+    public override int PlayGame() {
+        return 100;
+    }
+}
